% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{basic_statistics}
\alias{basic_statistics}
\title{Calculate basic network statistics for a network and print to screen.}
\usage{
basic_statistics(
  g,
  name = NULL,
  directed = FALSE,
  unconnected = TRUE,
  weights = NULL,
  verbose = FALSE
)
}
\arguments{
\item{g}{An igraph object.}

\item{name}{A name for the network (optional).}

\item{directed}{Logical, whether directed or undirected paths are to be
considered. This is ignored for undirected graphs.}

\item{unconnected}{Logical, what to do if the graph is unconnected. If
FALSE, the function will return a number that is one larger the largest
possible diameter, which is always the number of vertices. If TRUE, the
diameters of the connected components will be calculated and the largest one
will be returned.}

\item{weights}{Optional positive weight vector for calculating weighted
distances. If the graph has a weight edge attribute, then this is used by
default.}

\item{verbose}{Print progress to console.}
}
\description{
Calculate basic network statistics for a network and print to screen.
}
\seealso{
\code{\link{igraph::vcount}}, \code{\link{igraph::ecount}},
\code{\link{igraph::diameter}}.
}
