% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{basic_statistics_multiplex}
\alias{basic_statistics_multiplex}
\title{Calculate basic network statistics for each layer of a multiplex
network and print to screen.}
\usage{
basic_statistics_multiplex(mpo, verbose = FALSE)
}
\arguments{
\item{mpo}{A multiplex object.}

\item{verbose}{Print progress to console.}
}
\description{
Calculate basic network statistics for each layer of a multiplex
network and print to screen.
}
