% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{calculate_tau}
\alias{calculate_tau}
\title{Calculate Tau}
\usage{
calculate_tau(mpo, reference_network, verbose = FALSE)
}
\arguments{
\item{mpo}{Multiplex network}

\item{reference_network}{Reference network}

\item{verbose}{Print progress to console.}
}
\value{
Tau vector (numeric)
}
\description{
Calculate tau vector for the given multiplex
network based on the reference network.
}
