% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{compare_networks}
\alias{compare_networks}
\title{Compare two networks with the given metric.}
\usage{
compare_networks(g, h, metric = "overlap", verbose = FALSE)
}
\arguments{
\item{g}{Reference network.}

\item{h}{Network of interest.}

\item{metric}{Name of the metric to use for comparison.}

\item{verbose}{Print progress to console.}
}
\value{
Score
}
\description{
Compare two networks with the given metric.
}
