% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{exclusivity}
\alias{exclusivity}
\title{Exclusivity}
\usage{
exclusivity(mpo, verbose = FALSE)
}
\arguments{
\item{mpo}{A multiplex object}

\item{verbose}{Print progress to console.}
}
\description{
Exclusivity is proportion of all edges in the multiplex that
are found in only one layer. Here we show how many edges are found in 1
layer, 2 layers, 3 layers etc...
}
