% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{get_name}
\alias{get_name}
\title{Get the network name.}
\usage{
get_name(g, default = "<G>")
}
\arguments{
\item{g}{An igraph network object.}

\item{default}{The default network name.}
}
\value{
A string.
}
\description{
Get the network name.
}
