% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{jaccard_score_edges}
\alias{jaccard_score_edges}
\title{Jaccard similarity coefficient of edges for two networks.}
\usage{
jaccard_score_edges(g, h, verbose = FALSE)
}
\arguments{
\item{g}{Reference network.}

\item{h}{Network of interest.}

\item{verbose}{Print progress to console.}
}
\value{
Jaccard similarity coefficient.
}
\description{
Jaccard similarity coefficient of edges for two networks.
}
