% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{load_flist}
\alias{load_flist}
\title{Load flist}
\usage{
load_flist(path_to_flist, verbose = FALSE)
}
\arguments{
\item{path_to_flist}{A path to an flist file.}

\item{verbose}{Print progress to console.}
}
\value{
A data.table
}
\description{
\code{load_flist} loads an flist file from a given path.
An flist file is a tab-delimited file that describes a
multiplex network. It should have the following three columns:
}
\details{
"path to file" "short name of network" 'group"

'groups' are either 1, 2, or 3.  All 1's will form one multiplex network
(e.g. gene-to-gene), All 2's will form a separate multiplex network (e.g.
disease-to-disease), And all 3's will be used to join the 1's and 2's
together (e.g. gene-to-disease) You don't have to have both 1's and 2's. But
if you do have 1's and 2's, you SHOULD have at least one 3 to join them up.
}
