% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_network}
\alias{load_network}
\title{Load Network}
\usage{
load_network(
  path_to_edgelist,
  type = NULL,
  name = NULL,
  col_names = c("from", "to", "weight"),
  select = 1:3,
  header = "auto",
  directed = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{path_to_edgelist}{A path to a network file in edgelist format.
Expected to have two columns: 'from', 'to', and 'weight'.}

\item{type}{The edge type.}

\item{name}{The name of the network.}

\item{col_names}{The column names to use for the edgelist.}

\item{select}{The columns to select from the edgelist.}

\item{directed}{Whether the network is directed.}

\item{verbose}{Print progress to console.}
}
\value{
An igraph network object.
}
\description{
Loads an igraph network from a filepath to an edge list.
}
