% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{make_dummy_multiplex}
\alias{make_dummy_multiplex}
\title{Make a \emph{dummy} multiplex network object.}
\usage{
make_dummy_multiplex(flist_or_path, verbose = FALSE)
}
\arguments{
\item{flist_or_path}{A data.table or path to an flist file.}

\item{verbose}{Print progress to console.}
}
\value{
An multiplex network object.
}
\description{
Create a multiplex network object from the provided flist. This
is a '\emph{dummy}' multiplex network object because it is simply a list of
network layers with a single extra attribute \code{Number_of_Layers}.
}
