% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_network_aggregation.R
\name{merged_with_all_edges}
\alias{merged_with_all_edges}
\title{Merged with All Edges}
\usage{
merged_with_all_edges(mpo, verbose = FALSE)
}
\arguments{
\item{mpo}{A multiplex network object.}

\item{verbose}{Print progress to console.}
}
\value{
An igraph network object.
}
\description{
Merge down all layers in a multiplex object, but don't
aggregate the edges (i.e. keep all edges). This function can take a dummy
multiplex or a real multiplex.
}
