% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_network_aggregation.R
\name{merged_with_edgecounts}
\alias{merged_with_edgecounts}
\title{Merged With Edgecounts}
\usage{
merged_with_edgecounts(mpo, inv = FALSE, verbose = FALSE)
}
\arguments{
\item{mpo}{A multiplex network object. Edge weights are the sum of the
weights of the edges in the multiplex network object.}

\item{inv}{Set the edge weight to the reciprocal of the sum of the weights.}

\item{verbose}{Print progress to console.}
}
\value{
An igraph network object.
}
\description{
Merge down all layers and aggregate multi-edges into one edge
where edge weight is the number of layers in which the two nodes are
connected.
}
