% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{overlap_many_pairwise}
\alias{overlap_many_pairwise}
\title{Calculate overlap scores between all layers of a multiplex network.}
\usage{
overlap_many_pairwise(mpo, metric = "overlap", verbose = FALSE)
}
\arguments{
\item{mpo}{A multiplex object.}

\item{metric}{Metric to use for comparison.}

\item{verbose}{Print progress to console.}
}
\value{
numeric vector
}
\description{
Calculate overlap scores between all layers of a multiplex network.
}
