% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{overlap_many_vs_reference}
\alias{overlap_many_vs_reference}
\title{Calculate Many Vs. Reference}
\usage{
overlap_many_vs_reference(
  mpo,
  reference_network,
  metric = "overlap",
  verbose = FALSE
)
}
\arguments{
\item{mpo}{multiplex network}

\item{reference_network}{reference network}

\item{metric}{metric to use for calculating overlap score}

\item{verbose}{Print progress to console.}
}
\value{
numeric vector
}
\description{
Calculate overlap scores between a multiplex network and a reference network.
}
