% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{overlap_pair}
\alias{overlap_pair}
\title{Compare two networks with the given metric.}
\usage{
overlap_pair(...)
}
\arguments{
\item{g}{Reference network.}

\item{h}{Network of interest.}

\item{metric}{Metric to use for comparison.}

\item{verbose}{Print progress to console.}
}
\value{
Score
}
\description{
Compare two networks with the given metric.
}
