% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{overlap_score}
\alias{overlap_score}
\title{Sum of edge weights of the intersection divided by the number of
edges in the network of interest.}
\usage{
overlap_score(g, h, verbose = FALSE)
}
\arguments{
\item{g}{Reference network.}

\item{h}{Network of interest.}

\item{verbose}{Print progress to console.}
}
\value{
Overlap score.
}
\description{
Sum of edge weights of the intersection divided by the number of
edges in the network of interest.
}
