% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netstats.R
\name{RWR_netstats}
\alias{RWR_netstats}
\title{Command-line interface for RWR_netstats.R}
\usage{
RWR_netstats(
  data = NULL,
  flist = NULL,
  network_1 = NULL,
  network_2 = NULL,
  basic_statistics = F,
  overlap_sim_multiplex = F,
  overlap_sim_multiplex_layer = F,
  overlap_sim_multiplex_layer_jaccard = F,
  overlap_sim_layer_layer = F,
  overlap_score = F,
  calculate_tau_for_mpo = F,
  merged_with_all_edges = F,
  merged_with_edgecounts = F,
  calculate_exclusivity_for_mpo = F,
  verbose = F
)
}
\arguments{
\item{data}{The filepath to an mpo object.}

\item{flist}{An flist. Currently creates "faux mpo"
object
for fast statistical inferences.}

\item{network_1}{A path to an edgelist. Used for basic
statistics, overlap_sim_multiplex_layer,
overlap_pair, and calculate tau}

\item{network_2}{A path to an edgelist. Used for basic
statistics and overlap_pair.}

\item{basic_statistics}{A boolean denoting a return for basic
statistics concerning supplied networks,
or flists.}

\item{overlap_sim_multiplex}{A boolean denoting a return of jaccard
similarity metrics for the supplied
multiplex}

\item{overlap_sim_multiplex_layer}{A boolean denoting a return of the
calculated edge weight overlap between
a multiplex network and a reference
network (supplied as "network_1")}

\item{overlap_sim_layer_layer}{A boolean denoting a return of jaccard
and edge weight overlap between two
supplied networks (network_1 and
network_2)}

\item{overlap_score}{A boolean denoting a return of a matrix
of overlap scores between all layers
of the multiplex.}

\item{merged_with_all_edges}{A boolean denoting a return of a merged
down multiplex network along with
network edge counts and vertex counts.}

\item{merged_with_edgecounts}{A boolean denoting a return of a merged
down multiplex, but simplified with
edge weights denoting the total number
of layers in which that edge existed.}

\item{verbose}{A boolean denoting the verbosity of
output}

\item{calculate_tau}{A boolean denoting a return of the
distribution of "tau", with respect
to the network layers,  calculated via
edge overlap weight / total edgeweight
multipled by the total number of layers}

\item{exclusivity}{A boolean denoting a return of total
percentage of edges that exist within
all n layers of the multiplex.}
}
\value{
A list of data frames, igraph networks, and vectors denoting various
parameter defined statistics.
}
\description{
Command-line interface for RWR_netstats.R
}
