% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_shortestpaths.R
\name{RWR_ShortestPaths}
\alias{RWR_ShortestPaths}
\title{RWR Shortest Paths}
\usage{
RWR_ShortestPaths(
  multiplex_filepath = NULL,
  source_geneset = NULL,
  target_geneset = NULL,
  outdir = ".",
  out_path = NULL,
  threads = 1,
  cyto = FALSE,
  verbose = FALSE,
  write_to_file = FALSE
)
}
\arguments{
\item{multiplex_filepath}{Path to the .Rdata file for your combo of
underlying functional networks. This file is produced by
RWR_make_MHobject.R}

\item{source_geneset}{Path to the gene set file. If given with
--target_geneset, find shortest paths between genes in --source_geneset and
--target_geneset. Otherwise, find shortest paths among genes in
--source_geneset only. It must have the following cols without heading:
{<}setid{>} {<}gene{>}.}

\item{target_geneset}{Path to the second geneset file. If it is not provided
then the shortest paths will be calculated between genes in source_geneset.}

\item{outdir}{Full path to the output file directory. Two output files will
be generated with different suffixes.  Default "."}

\item{out_path}{Specify the full path for output. Ignore --outdir and
--modname and use this instead.}

\item{threads}{Number of threads to use.}

\item{cyto}{Include this parameter if you wish to see the shortest paths in
Cytoscape. Cytoscape must already be running first!  If your geneset(s) are
large (e.g. 100+) then loading all the shortest paths fully into cytoscape
can take a while. Default FALSE}

\item{verbose}{Verbose mode.  Default FALSE}

\item{write_to_file}{Also write the results to a file.  Default FALSE}
}
\value{
Returns a data frame.
}
\description{
\code{RWR_ShortestPaths} Find shortest paths between genes in the given gene
sets. These will be output as a table of edges. There are two ways to use
this:
\enumerate{
\item provide one geneset (g) and you will get all shortest paths between genes
in g.
\item provide two genesets (g and p) and you will get the shortest paths
between genes in g and genes in p.
}
}
\examples{

# An example of Running RWR_ShortestPaths

extdata.dir <- system.file("example_data", package="RWRtoolkit")
multiplex_object_filepath <- paste(extdata.dir,
                                   '/string_interactions.Rdata',
                                   sep='')
geneset1_filepath <- paste(extdata.dir, '/geneset1.tsv', sep='')
geneset2_filepath <- paste(extdata.dir, '/geneset2.tsv', sep='')
outdir <- paste(extdata.dir, '/out/rwr_shortestpath', sep='')

rwr_shortest_path_output <- RWR_ShortestPaths(data=multiplex_object_filepath,
                                            source_geneset=geneset1_filepath,
                                            target_geneset=geneset2_filepath,
                                            write_to_file=TRUE,
                                            outdir=outdir)



}
