% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR_netscore.R
\name{RWR_netscore}
\alias{RWR_netscore}
\title{RWR netscore}
\usage{
RWR_netscore(
  gold = NULL,
  network = NULL,
  reference_geneset = NULL,
  perm = 0,
  threads = NULL,
  outdir = ".",
  write_to_file = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{gold}{Full path to the gold network file. This is the network that
contains known truth edges.  Input file is a table with first 3 cols {<}GeneA{>}
{<}GeneB{>} {<}score{>} column names don't matter.}

\item{network}{Full path to your network to be evaluated. This is a table
with first 3 cols {<}GeneA{>} {<}GeneB{>} {<}score{>} column names don't matter.}

\item{reference_geneset}{Only required if using \code{perm}. Full path to a file
containing all gene/node labels that were used in the process of generating
the network provided in \code{network}.}

\item{perm}{Number of random permutations to use for estimating
significance. If you set this, you must also set \code{refgenes}.}

\item{threads}{Only set this is you want to do permutations.}

\item{outdir}{Output directory. Default "."}

\item{write_to_file}{Also write the result to a file. Default FALSE}

\item{verbose}{Verbose mode. Default FALSE}
}
\value{
Returns a table of metrics as a comparison between networks with parameters:
network:  Original Network input.
gold:   Gold Network comparison.
network_nodes : Number of connected nodes within network (i.e. singleton nodes will not be counted)
gold_nodes:     Number of nodes within the gold network
network_edges:  Total number of edges within the network.
gold_edges:     Total number of edges within the gold network
nodes_in_common:  Number of nodes existing within both input and gold networks.
intersecting_edges: Number of edges existing within both input and gold networks.
intersect_perc: Percentage of intersecting nodes
intersect_score:  @dkainer
score_per_network_edge: @dkainer
}
\description{
\code{RWR_netscore} This tool performs a network intersect between your input
network (--network) and a gold truth network (--gold). It scores the
strength of that intersect with multiple metrics.
}
\examples{

extdata.dir <- system.file("example_data", package="RWRtoolkit")
multiplex_object_filepath <- paste(extdata.dir, '/string_interactions.Rdata', sep='')
geneset1_filepath <- paste(extdata.dir, '/geneset1.tsv', sep='')
geneset2_filepath <- paste(extdata.dir, '/geneset2.tsv', sep='')
outdir <- paste(extdata.dir, '/out/rwr_netscore', sep='') 

# An example of Running RWR_ShortestPaths
netscore_table <- RWRtoolkit::RWR_ShortestPaths(data=multiplex_object_filepath,
                                              source_geneset=geneset1_filepath,
                                              target_geneset=geneset2_filepath,
                                              write_to_file=TRUE,
                                              outdir=outdir)



}
